requirements.txt
absl-py==0.11.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
autocorrect==2.2.2
backcall==0.2.0
base58==2.0.1
bayesian-optimization==1.2.0
bert-for-tf2==0.14.7
bleach==3.2.1
blinker==1.4
boto3==1.16.23
botocore==1.19.23
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorful==0.5.4
comet-ml==3.2.5
configobj==5.0.6
contractions==0.0.25
decorator==4.4.2
defusedxml==0.6.0
dulwich==0.20.13
emot==2.1
entrypoints==0.3
enum-compat==0.0.3
everett==1.0.3
fastapi==0.61.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras-Preprocessing==1.1.2
llvmlite==0.34.0
Markdown==3.3.3
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.9
nltk==3.5
notebook==6.1.5
numba==0.51.2
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
params-flow==0.8.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.12.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
py==1.9.0
py-params==0.9.7
pyahocorasick==1.4.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.2
pydeck==0.5.0
pydot==1.4.1
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.4
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
starlette==0.13.6
streamlit==0.71.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-hub==0.10.0
tensorflow-text==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
textsearch==0.0.17
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.53.0
traitlets==5.0.5
typeguard==2.10.0
tzlocal==2.1
umap-learn==0.4.6
Unidecode==1.1.1
urllib3==1.26.2
validators==0.18.1
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.1
